+++ Templates/Altruist.Templates/content/altruist-server/Program.cs
@@ -1,0 +1,3 @@
+using Altruist;
+
+await AltruistApplication.Run(args);